put email verification in newUserSubscription, need to test if it successfully sends
