+++ server/src/services/LeadurServices.js
@@ -29,19 +29,6 @@ const LeadurServices = {
                 reject(err);
               } else {
                 newUserSubscription.add(newUser);
-
-                const emailToken = jwt.sign(
-                  { email },
-                  process.env.JWT_RESET_TOKEN,
-                );
-                const url = process.env.CLIENT_BASE_URL + '/verify-user-email/' + email + '/' + emailToken;
-                const result = await EmailServices.sendSimpleEmail(
-                  [email],
-                  '',
-                  'Please use this URL to confirm your email: ' + url,
-                  'F!rosh Email Confirmation',
-                  'tech@orientation.skule.ca',
-                );
                 resolve(newUser);
               }
             },
