+++ server/src/subscribers/newUserSubscription.js
@@ -1,29 +1,40 @@
 const Queue = require('bull');
 const EmailServices = require('../services/EmailServices');
+const jwt = require('jsonwebtoken');
 
 const newUserSubscription = new Queue('newUser', {
   redis: { port: process.env.REDIS_PORT, host: 'redis', password: process.env.REDIS_PASSWORD },
 });
 
-newUserSubscription.process((job, done) => {
-  console.log(`New User created!`);
-  console.log(job.data);
+newUserSubscription.process(async (job, done) => {
   try {
-    // sending user creation email
-    const result = EmailServices.sendTemplateEmail(
-      {},
-      'signup_confirmation',
+    // sending user email verification link
+    const emailToken = jwt.sign(
       [job.data.email],
-      'tech@orientation.skule.ca',
+      process.env.JWT_RESET_TOKEN,
     );
-
-    result.then((response) => {
-      console.log(response);
+    const url = process.env.CLIENT_BASE_URL + '/verify-user-email/' + email + '/' + emailToken;
+    await EmailServices.sendSimpleEmail(
+       [job.data.email],
+       '',
+       'Please use this URL to confirm your email: ' + url,
+       'F!rosh Email Confirmation',
+       'tech@orientation.skule.ca',
+    ).then(() => {
+      try {
+        // sending user creation email
+        const result = EmailServices.sendTemplateEmail(
+          {},
+          'signup_confirmation',
+          [job.data.email],
+          'tech@orientation.skule.ca',
+        ); 
+      } catch (error) {
+        done(error);
+      }
     });
-
     done();
   } catch (error) {
-    console.log(error);
     done(error);
   }
 });
